added routers; removed server selection logic; popups logic changed
