+++ package.json
@@ -29,6 +29,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-infinite-scroll-component": "^6.1.0",
+    "react-router": "^7.1.5",
     "start": "^5.1.0",
     "styled-components": "^5.3.11",
     "typescript": "^5.4.5",
