+++ src/index.tsx
@@ -2,16 +2,16 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
-import {ServiceProvider} from './services/ServiceProvider';
+import {Providers} from './Providers';
 
 const rootElement = document.getElementById('root');
 if (rootElement) {
     const root = ReactDOM.createRoot(rootElement);
     root.render(
         <React.StrictMode>
-            <ServiceProvider>
+            <Providers>
                 <App />
-            </ServiceProvider>
+            </Providers>
         </React.StrictMode>
     );
 }
